requirements.txt
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.1.0
pruna-ai>=0.1.0
numpy>=1.20.0
tqdm>=4.64.0
pillow>=9.0.0
matplotlib>=3.7.0
pyyaml>=6.0 